Added potential goalpost sprite in case we decide to add one.
should preferably be treated as a separate fourth tile, moving sprite to
separate memory. If so, make it's background fit in with the environment
(make it sit on ice/ground, or something). May alternatively be seen a s
a pink seal, if that makes more sense. Also, should preferably be
arranged so that Pengucock sprite can only reach goalpost from below.